Add files via upload
I recently developed a static product launch page titled "Smartify," leveraging my skills in Java, HTML, and CSS. This project was deployed on a Tomcat server (version 10.1.24) and developed using Eclipse as the integrated development environment. The Smartify page is designed to highlight the features and benefits of the product effectively, providing users with an engaging and informative browsing experience.

The development process involved creating a clean and responsive design using HTML and CSS, ensuring that the page is visually appealing and user-friendly across various devices and screen sizes. Utilizing Java for backend support allowed for a robust and scalable deployment on the Tomcat server. This project not only showcases my technical expertise in web development but also demonstrates my ability to create and manage static web pages efficiently.

Through this project, I have reinforced my skills in Java-based web development and my proficiency in creating visually appealing and functional web interfaces. The Smartify launch page is a testament to my dedication to delivering high-quality web solutions.